polish the response screen
o.fullContrastResponseAlternatives
It was important to move the computation of new CLUT to be BEFORE we computed the response screen.
o.signalIsBinary
Treat binary signals as a special case, imaged using o.gray1 for LMean, whereas the general case of grayscale and color images use o.gray for LMean.
o.printGrayLuminance