fixed ModelObserver
the trick is to pass the “locations” struct that was saved in the movie, as a way of grabbing the signal in noise. it’s complete and reliable. it is important to skip any pre-frames, because they are blank.

Also enhanced the code that generates a filename. now, if it detects a conflict it advances time, but 1 s at a time, until it finds an unused file name. this became an issue because the computer ideal does a whole block in less than a second.
added several checks for proper arguments to ObserverModel.