polish interleaving. oo(1).quitExperiment o.signalImagesCacheCode
Setting o.quitExperiment now also sets oo(1).quitExperiment, which is returned when NoiseDiscrimination2 returns.
if you provide a o.signalImagesCacheCode indicating the same signal as another condition, then, when the codes match, we reuse the already loaded signal. This is a hugentimesaver when loading many conditions with the same signals.