greatly reduced usage of “o”.
it would be good to finish getting rid of it. “o” is now an attractive nuissance.. The code looks right and may behave approximately right, but “o” tends to be stale which can create bad problems that are easily overllooked. If i got rid of o, then MATLAB would flag any attempt to use “o”.