Debugged EstimateNeq
It was making terrible fits. mincon with the default algorithm often gave wrong answers. Switching to the ‘sqp’ algorithm helped a lot. We now try an unconstrained fit first, which is more reliable, and only if the answer is out of bounds do we try the constrained fit. Also, when the slope is slightly negative it’s important to seed with large Neq. I think the fitting is now reliable.