new
remove ff (which might not be valid) from CloseWindows…
so use fprintf instead of ffprintf.
In several places change
if isempty(o.window)
to
if isempty(o.window) || ismember(o.observer,o.algorithmicObservers)
to cope with situation in which observer is algorithmic, yet window is already open.
Change “X” to X in on-screen messages.
Move warnings about debug mode slightlly later in printout for easier reading.
Rename winfo to windowInfo.
If we detect that “window” is invalid we set it to [] and set isFirstBlock=true;