polish Omkar’s code to display uncertainty
also changed some unbreakable spaces to spaces, because the unbreakable space is not ASCII, and was not preserved on Omakar’s computer.  We should figure out how to mark the MATLAB m file as unicode so that 16 bit characters are conserved.

There is still a plus/minus character that will not be conserved on Omkar’s computer until we fix the unicode issue.